refactor: Move externals into seperate packages
