+++ consumatio/gateways/popular_gateways/popular_tv_to_dict.py
@@ -1,5 +1,5 @@
 from consumatio.entities.tv import TV
-from consumatio.external.models import *
+from consumatio.external.db.models import *
 from sqlalchemy import text
