+++ consumatio/gateways/search_gateways/search_result_to_dict.py
@@ -1,6 +1,6 @@
 from consumatio.entities.movie import Movie
 from consumatio.entities.tv import TV
-from consumatio.external.models import *
+from consumatio.external.db.models import *
 from sqlalchemy import text
