+++ consumatio/usecases/get_movie.py
@@ -1,5 +1,5 @@
 from consumatio.entities.movie import Movie
-from consumatio.external.models import *
+from consumatio.external.db.models import *
 
 
 def get_movie(external_id: str, tmdb: object, code: int, country: str) -> dict:
