+++ consumatio/usecases/set_rating.py
@@ -16,7 +16,7 @@ def set_rating(database: object, external_id: str, media: str, code: int,
     user_id = database.get_user_id(external_id)
 
     if not database.media_data_exists(user_id, media, code):
-        database.media_Data(user_id, media, code)
+        database.media_Data(user_id, media, code, tv_code=None)
 
     database.rating(user_id, media, code, rating)
     return {"status": True}
